refactor(api): split crawlers.js controller into per-responsibility modules (#599)
crawlers.js had grown to 657 lines (audit finding C1). Split it by
responsibility into focused routers under routes/crawlers/:

  admin.js       — /api/admin/crawlers[...] (Entra ID auth) -> adminCrawlersRouter
  selfService.js — /api/crawlers/* (API-key auth: self-service, worker job-claim
                   protocol, delta-token persistence) -> selfServiceCrawlersRouter
  shared.js      — useSql + the API-key generateApiKey / hashKey helpers

app.js mounts the two routers separately (different auth) and the
OpenAPI-drift guard introspects both, so crawlers.js is now a barrel that
re-exports the two named routers exactly as before (no default export).
Pure code move — each route module also imports `crypto` directly (the
handlers call crypto.randomBytes for the key salt; `crypto` is otherwise
the global Web Crypto object, which lacks randomBytes).

Grandfathered cognitive entries relocated (no increase): [16,16] ->
admin [16] + selfService [16] (no cyclomatic units over threshold).
Extended crawlers.routes.test.js with admin (audit/patch/reset) and
self-service (rotate, job-progress, claim, mark-delta-mode, delta-tokens,
phases/complete/fail) happy paths — lifting part2's changed-line diff
coverage to 84%. Full API suite green (1786 tests); complexity + jscpd
(0 net-new clones) + OpenAPI-drift gates pass.
