+++ .ci/cognitive-baseline.json
@@ -75,8 +75,10 @@
     "app/api/src/routes/contexts/crud.js": [
       40
     ],
-    "app/api/src/routes/crawlers.js": [
-      16,
+    "app/api/src/routes/crawlers/admin.js": [
+      16
+    ],
+    "app/api/src/routes/crawlers/selfService.js": [
       16
     ],
     "app/api/src/routes/details/accessPackage.js": [
